requirements.txt
absl-py==2.3.1
agate==1.9.1
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==14.0
contourpy==1.3.2
cycler==0.12.1
daff==1.4.2
dagster==1.11.2
dagster-pipes==1.11.2
dagster_shared==1.11.2
dbt-adapters==1.16.1
dbt-common==1.26.0
dbt-core==1.10.4
dbt-extractor==0.6.0
dbt-protos==1.0.337
dbt-semantic-interfaces==0.9.0
deepdiff==7.0.1
docstring_parser==0.16
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.73.1
grpcio-health-checking==1.71.2
h11==0.16.0
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.7
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
leather==0.4.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
parsedatetime==2.6
pathspec==0.12.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.6.1
pybboxes==0.1.6
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
roboflow==1.2.1
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
sahi==0.11.30
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
Telethon==1.40.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
text-unidecode==1.3
thop==0.1.1.post2209072238
tomli==2.2.1
tomlkit==0.13.3
toposort==1.10
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.166
ultralytics-thop==2.0.14
universal_pathlib==0.2.6
urllib3==2.5.0
uvicorn==0.35.0
watchdog==5.0.3
Werkzeug==3.1.3
yolov5==7.0.14
yolov8==0.0.2
zipp==3.23.0
